+++ app/src/main/java/org/breezyweather/sources/pirateweather/PirateWeatherResultConverter.kt
@@ -52,20 +52,20 @@ internal fun getCurrent(
         weatherText = result.summary,
         weatherCode = getWeatherCode(result.icon),
         temperature = TemperatureWrapper(
-            temperature = validate(result.temperature),
-            feelsLike = validate(result.apparentTemperature)
+            temperature = result.temperature,
+            feelsLike = result.apparentTemperature
         ),
         wind = Wind(
-            degree = validate(result.windBearing),
-            speed = validate(result.windSpeed),
-            gusts = validate(result.windGust)
+            degree = result.windBearing,
+            speed = result.windSpeed,
+            gusts = result.windGust
         ),
-        uV = UV(index = validate(result.uvIndex)),
-        relativeHumidity = validate(result.humidity)?.times(100),
-        dewPoint = validate(result.dewPoint),
-        pressure = validate(result.pressure),
-        cloudCover = validate(result.cloudCover)?.times(100)?.roundToInt(),
-        visibility = validate(result.visibility)?.times(1000),
+        uV = UV(index = result.uvIndex),
+        relativeHumidity = result.humidity?.times(100),
+        dewPoint = result.dewPoint,
+        pressure = result.pressure,
+        cloudCover = result.cloudCover?.times(100)?.roundToInt(),
+        visibility = result.visibility?.times(1000),
         dailyForecast = dailySummary,
         hourlyForecast = hourlySummary
     )
@@ -81,8 +81,8 @@ internal fun getDailyForecast(
                 weatherPhase = result.summary,
                 weatherCode = getWeatherCode(result.icon),
                 temperature = TemperatureWrapper(
-                    temperature = validate(result.temperatureHigh),
-                    feelsLike = validate(result.apparentTemperatureHigh)
+                    temperature = result.temperatureHigh,
+                    feelsLike = result.apparentTemperatureHigh
                 )
             ),
             night = HalfDayWrapper(
@@ -91,11 +91,11 @@ internal fun getDailyForecast(
                 // temperatureLow/High are always forward-looking
                 // See https://docs.pirateweather.net/en/latest/API/#temperaturelow
                 temperature = TemperatureWrapper(
-                    temperature = validate(result.temperatureLow),
-                    feelsLike = validate(result.apparentTemperatureLow)
+                    temperature = result.temperatureLow,
+                    feelsLike = result.apparentTemperatureLow
                 )
             ),
-            uV = UV(index = validate(result.uvIndex))
+            uV = UV(index = result.uvIndex)
         )
     }
 }
@@ -112,32 +112,32 @@ internal fun getHourlyForecast(
             weatherText = result.summary,
             weatherCode = getWeatherCode(result.icon),
             temperature = TemperatureWrapper(
-                temperature = validate(result.temperature),
-                feelsLike = validate(result.apparentTemperature)
+                temperature = result.temperature,
+                feelsLike = result.apparentTemperature
             ),
             // see https://docs.pirateweather.net/en/latest/API/#precipaccumulation
             precipitation = Precipitation(
-                total = validate(result.precipAccumulation)?.times(10),
-                rain = validate(result.liquidAccumulation)?.times(10),
-                snow = validate(result.snowAccumulation)?.times(10),
-                ice = validate(result.iceAccumulation)?.times(10)
+                total = result.precipAccumulation?.times(10),
+                rain = result.liquidAccumulation?.times(10),
+                snow = result.snowAccumulation?.times(10),
+                ice = result.iceAccumulation?.times(10)
             ),
             precipitationProbability = PrecipitationProbability(
-                total = validate(result.precipProbability)?.times(100)
+                total = result.precipProbability?.times(100)
             ),
             wind = Wind(
-                degree = validate(result.windBearing),
-                speed = validate(result.windSpeed),
-                gusts = validate(result.windGust)
+                degree = result.windBearing,
+                speed = result.windSpeed,
+                gusts = result.windGust
             ),
             uV = UV(
-                index = validate(result.uvIndex)
+                index = result.uvIndex
             ),
-            relativeHumidity = validate(result.humidity)?.times(100),
-            dewPoint = validate(result.dewPoint),
-            pressure = validate(result.pressure),
-            cloudCover = validate(result.cloudCover)?.times(100)?.roundToInt(),
-            visibility = validate(result.visibility)?.times(1000)
+            relativeHumidity = result.humidity?.times(100),
+            dewPoint = result.dewPoint,
+            pressure = result.pressure,
+            cloudCover = result.cloudCover?.times(100)?.roundToInt(),
+            visibility = result.visibility?.times(1000)
         )
     }
 }
@@ -158,7 +158,7 @@ internal fun getMinutelyForecast(minutelyResult: List<PirateWeatherMinutely>?): 
                 } else {
                     ((minutelyForecast.time - minutelyResult[i - 1].time) / 60).toDouble().roundToInt()
                 },
-                precipitationIntensity = validate(minutelyForecast.precipIntensity)
+                precipitationIntensity = minutelyForecast.precipIntensity
             )
         )
     }
@@ -203,12 +203,5 @@ private fun getWeatherCode(icon: String?): WeatherCode? {
         "partly-cloudy-day", "partly-cloudy-night" -> WeatherCode.PARTLY_CLOUDY
         "cloudy" -> WeatherCode.CLOUDY
         else -> null
-    }
-}
-
-private fun validate(input: Double?): Double? {
-    return when (input) {
-        -999.0 -> null
-        else -> input
     }
 }